omsBackend/requirements.txt
amqp==2.3.2
billiard==3.5.0.4
celery==4.2.1
certifi==2018.8.24
chardet==3.0.4
Django==2.1.1
django-cors-headers==2.4.0
django-filter==2.0.0
django-rest-auth==0.9.3
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
dry-rest-permissions==0.1.10
idna==2.7
kombu==4.2.1
Pillow==5.2.0
pyasn1==0.4.4
PyJWT==1.6.4
pytz==2018.5
requests==2.19.1
six==1.11.0
urllib3==1.23
vine==1.1.4
